Version 3.0 - AI & Data Gathering
Added AI that can play the game - play based on bomb probabilities in any available cells.

In addition, the ability to collect data on the amount of wins and losses of a specific board size and amount of bombs. This will be useful in future data visualization analysis and to monitor the improvement of various AI algorithms. Also, it will help to identify the sweet spot for guaranteed win, guaranteed loss and the probabilities of winning or losing in between.

Hint, there seems to be a repeating pattern.